Create CurrencyView which shows title, currency and amount and also look differently based on the fact is it selected or not - use it together with CurrencyConverterView so app matches the design more

+++ TransferGO/TransferGO/CurrencyConverter/CurrencyConverterInfo.swift
@@ -10,5 +10,5 @@ import Foundation
 struct CurrencyConverterInfo {
     var fromCountry: Country?
     var toCountry: Country?
-    var fromAmount: Int?
+    var fromAmount: Double?
 }
